Continued refactoring code incl. time zone parsing
