+++ app.py
@@ -46,29 +46,21 @@ def load_config():
 
     if "time_zone" in config_dict["settings"] and config_dict["settings"]["time_zone"]:
         time_zone = config_dict["settings"]["time_zone"]
-        app_time_zone, time_zone_name = utility.time_zone_parser(time_zone)
+        time_zone_object, time_zone_string = utility.time_zone_parser(time_zone)
 
-        config_dict["settings"]["app_time_zone"] = app_time_zone
-        config_dict["settings"]["time_zone"] = time_zone_name
+        config_dict["settings"]["app_time_zone"] = time_zone_object
+        config_dict["settings"]["time_zone"] = time_zone_string
+        config_dict["database"]["time_zone"] = time_zone_string
     else:
         config_dict["settings"]["app_time_zone"] = pytz.timezone("UTC")
         config_dict["settings"]["time_zone"] = "UTC"
+        config_dict["database"]["time_zone"] = "UTC"
 
     return config_dict
 
 #endregion
 
 #region Common Functions
-def generate_date_time_stamp(time_zone: pytz.timezone = pytz.timezone("UTC")):
-    """Generate a current date/timestamp string"""
-    now = datetime.now(time_zone)
-    return now.strftime("%Y-%m-%d %H:%M:%S %Z")
-
-def current_year(time_zone: pytz.timezone = pytz.timezone("UTC")):
-    """Return the current year"""
-    now = datetime.now(time_zone)
-    return now.strftime("%Y")
-
 def retrieve_show_dates(reverse_order: bool = False):
     """Retrieve a list of available show dates"""
     database_connection.reconnect()
@@ -95,18 +87,6 @@ def retrieve_show_years_months(reverse_order: bool = False):
         years_months.reverse()
 
     return years_months
-
-def date_string_to_date(**kwargs):
-    """Used to convert an ISO-style date string into a datetime object"""
-    if "date_string" in kwargs and kwargs["date_string"]:
-        try:
-            date_object = parser.parse(kwargs["date_string"])
-            return date_object
-
-        except ValueError:
-            return None
-
-    return None
 
 #endregion
 
@@ -602,7 +582,7 @@ def npr_show_redirect(show_date: Text):
     """Takes an ISO-like date string and redirects to the appropriate
     show page on NPR's website."""
     database_connection.reconnect()
-    show_date_object = date_string_to_date(date_string=show_date)
+    show_date_object = utility.date_string_to_date(date_string=show_date)
 
     if not show_date_object:
         return redirect(url_for("index"))
@@ -629,14 +609,16 @@ def npr_show_redirect(show_date: Text):
 
 #region Application Initialization
 config = load_config()
+app_time_zone = config["settings"]["app_time_zone"]
+time_zone_name = config["settings"]["time_zone"]
 app.jinja_env.globals["app_version"] = APP_VERSION
 app.jinja_env.globals["current_date"] = date.today()
-app.jinja_env.globals["date_string_to_date"] = date_string_to_date
+app.jinja_env.globals["date_string_to_date"] = utility.date_string_to_date
 app.jinja_env.globals["ga_property_code"] = config["settings"]["ga_property_code"]
-app.jinja_env.globals["current_year"] = current_year(config["settings"]["app_time_zone"])
+app.jinja_env.globals["current_year"] = utility.current_year(app_time_zone)
 app.jinja_env.globals["rank_map"] = dicts.PANELIST_RANKS
-app.jinja_env.globals["time_zone"] = config["settings"]["app_time_zone"]
-app.jinja_env.globals["rendered_at"] = generate_date_time_stamp(config["settings"]["app_time_zone"])
+app.jinja_env.globals["time_zone"] = app_time_zone
+app.jinja_env.globals["rendered_at"] = utility.generate_date_time_stamp(app_time_zone)
 
 app.jinja_env.globals["api_url"] = config["settings"]["api_url"]
 app.jinja_env.globals["blog_url"] = config["settings"]["blog_url"]
@@ -644,8 +626,7 @@ app.jinja_env.globals["graphs_url"] = config["settings"]["graphs_url"]
 app.jinja_env.globals["reports_url"] = config["settings"]["reports_url"]
 app.jinja_env.globals["site_url"] = config["settings"]["site_url"]
 
-database_connection = mysql.connector.connect(**config["database"],
-                                              time_zone=config["settings"]["time_zone"])
+database_connection = mysql.connector.connect(**config["database"])
 database_connection.autocommit = True
 
 if __name__ == "__main__":
